+++ driver/web/apis_admin.go
@@ -224,7 +224,10 @@ func (h AdminApisHandler) refresh(l *logs.Log, r *http.Request, claims *tokenaut
 	}
 
 	accessToken := loginSession.AccessToken
-	refreshToken := loginSession.CurrentRefreshToken()
+	refreshToken, err := loginSession.CurrentRefreshToken()
+	if err != nil {
+		return l.HTTPResponseErrorData(logutils.StatusMissing, model.TypeRefreshToken, nil, err, http.StatusInternalServerError, false)
+	}
 
 	paramsRes, err := convert[Def.SharedResRefresh_Params](loginSession.Params)
 	if err != nil {